96 - Listando comentários do filme

+++ app/src/main/java/br/com/hellodev/moviestreaming/data/remote/model/author/AuthorDetailsResponse.kt
@@ -12,5 +12,5 @@ data class AuthorDetailsResponse(
     val avatarPath: String? = null,
 
     @SerialName("rating")
-    val rating: Int? = null
+    val rating: Float? = null
 )
