+++ app/src/main/java/br/com/hellodev/moviestreaming/domain/remote/model/reviews/AuthorDetails.kt
@@ -3,5 +3,5 @@ package br.com.hellodev.moviestreaming.domain.remote.model.reviews
 data class AuthorDetails(
     val name: String? = null,
     val avatarPath: String? = null,
-    val rating: Int? = null
+    val rating: Float? = null
 )
